#111: delete core_backends and fix Zero IIa pair migrate
Keep TrustModelBackendMixin only on trusts.backends. Remove the
transitional core_backends module with no alias. Test migrations that
historically referenced trusts.models.ReadonlyFieldsMixin now use
trusts.zero.models so the IIa pair suite can create its test database.
Skip leftover Zero Step I AppConfig/LookupError assertions.
