+++ trusts/test_issue103.py
@@ -1,20 +1,20 @@
-"""#103 / #111: generic ``trusts.core_backends`` boundary after cutover.
+"""#103 / #111: generic mixin lives only on ``trusts.backends``.
 
-Canonical mixin import plus a deprecated same-object alias on
-``trusts.backends``. Historical ``TrustModelBackend`` and TrustGroup
-SQL are gone from core. Runtime mixin behavior is unchanged.
+``trusts.core_backends`` is gone. Historical ``TrustModelBackend`` and
+TrustGroup SQL live only on ``trusts.zero.backends``.
 """
 
 import ast
+import importlib
 import inspect
 
 from django.test import SimpleTestCase
 
+from trusts.backends import TrustModelBackendMixin
 from trusts.core import PlanQueryCompiler
-from trusts.core_backends import TrustModelBackendMixin
 
 
-class CoreBackendsBoundaryTest(SimpleTestCase):
+class BackendsMixinBoundaryTest(SimpleTestCase):
     def test_canonical_mixin_import_and_plan_compiler(self):
         self.assertIsInstance(
             TrustModelBackendMixin.query_compiler, PlanQueryCompiler,
@@ -23,11 +23,9 @@ class CoreBackendsBoundaryTest(SimpleTestCase):
             TrustModelBackendMixin.query_compiler.historical_fallback,
         )
 
-    def test_backends_mixin_is_core_backends_mixin(self):
-        from trusts.backends import (
-            TrustModelBackendMixin as AliasMixin,
-        )
-        self.assertIs(AliasMixin, TrustModelBackendMixin)
+    def test_core_backends_module_is_gone(self):
+        with self.assertRaises(ModuleNotFoundError):
+            importlib.import_module('trusts.core_backends')
 
     def test_historical_backend_path_is_gone(self):
         import trusts.backends as backends_mod
@@ -39,16 +37,16 @@ class CoreBackendsBoundaryTest(SimpleTestCase):
         with self.assertRaises(ImportError):
             from trusts.backends import HistoricalGroupQueryCompiler  # noqa: F401
 
-    def test_core_backends_has_no_trustgroup_dependency(self):
-        import trusts.core_backends as core_backends
+    def test_backends_has_no_trustgroup_dependency(self):
+        import trusts.backends as backends
 
-        self.assertFalse(hasattr(core_backends, 'HistoricalGroupQueryCompiler'))
-        self.assertFalse(hasattr(core_backends, 'TrustModelBackend'))
-        self.assertFalse(hasattr(core_backends, 'historical_group_grant_exists'))
-        self.assertFalse(hasattr(core_backends, 'TrustGroup'))
-        self.assertNotIn('TrustGroup', core_backends.__dict__)
+        self.assertFalse(hasattr(backends, 'HistoricalGroupQueryCompiler'))
+        self.assertFalse(hasattr(backends, 'TrustModelBackend'))
+        self.assertFalse(hasattr(backends, 'historical_group_grant_exists'))
+        self.assertFalse(hasattr(backends, 'TrustGroup'))
+        self.assertNotIn('TrustGroup', backends.__dict__)
 
-        tree = ast.parse(inspect.getsource(core_backends))
+        tree = ast.parse(inspect.getsource(backends))
         imported = set()
         for node in ast.walk(tree):
             if isinstance(node, ast.ImportFrom):
